refactor(sync): import the shared streaming multipart module
The merge left two identical copies of the streaming multipart encoder:
sync/v1/_multipart.py from #241 and the product-neutral
assemblyai/_multipart.py the dictation client uses. Sync now imports the
shared one, including _Aborted, and its own copy is removed. No behaviour
change; the sync live tests import _STREAM_READ_SIZE from the shared path.
