+++ tests/unit/test_sync_live.py
@@ -17,8 +17,8 @@ import pytest
 from pytest_httpx import HTTPXMock
 
 import assemblyai as aai
+from assemblyai._multipart import _STREAM_READ_SIZE
 from assemblyai.sync.v1 import api, async_api
-from assemblyai.sync.v1._multipart import _STREAM_READ_SIZE
 
 aai.settings.api_key = "test"
